refactor: cleaner type & reuse
